// src/utils/enums.ts
enum SizeTypes {
  PIXEL = 'pixel',
  PERCENTEGE = 'percentage',
}

enum ResizeBarTypes {
  NORTH = 'north',
  SOUTH = 'south',
  EAST = 'east',
  WEST = 'west',
  NORTH_EAST = 'north-east',
  SOUTH_EAST = 'south-east',
  NORTH_WEST = 'north-west',
  SOUTH_WEST = 'south-west',
}

export { SizeTypes, ResizeBarTypes };